requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.6
pydantic>=2.6
pydantic-settings>=2.2
httpx>=0.27
pypdf>=4.0.0
pdfminer.six>=20231228
orjson>=3.10
slowapi>=0.1.8
aiofiles>=23.2.1
sqlmodel>=0.0.16
